% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadVars.R
\name{downloadVars}
\alias{downloadVars}
\title{downloadVars}
\usage{
downloadVars(vars = "all", ext = c(-180, 180, -90, 90))
}
\arguments{
\item{vars}{variables}

\item{ext}{extent}
}
\description{
downloadVars
}
