% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoTrend.R
\name{ecoTrend}
\alias{ecoTrend}
\title{ecoTrend}
\usage{
ecoTrend(occ, vars = "all", years = 1979:2013, res = 5, features = "lq")
}
\arguments{
\item{occ}{species occurrence coordinates (in this order: x, y or longitude, latitude), or an object from which coordinates can be extracted (e.g. a SpatVector of points)}

\item{vars}{variables}

\item{years}{year range}

\item{res}{spatial resolution (pixel size)}

\item{features}{feature classes for the Maxent model}
}
\description{
ecoTrend
}
