% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEverything.R
\name{getEverything}
\alias{getEverything}
\title{Get everything (package wrapper)}
\usage{
getEverything(occs, vars = "all", years = 1979:2013, region = NULL, res = 5)
}
\arguments{
\item{occs}{species occurrence coordinates (2 columns in this order: x, y or LONgitude, LATitude) in an object coercible to a data.frame (e.g. a data.frame, matrix, tibble, sf object or SpatVector of points), and in the same coordinate reference system as 'rasts'}

\item{vars}{character vector of the names of the variables to use - see varsAvailable()}

\item{years}{year range for the model time series; default 1979:2013}

\item{region}{optional SpatExtent or SpatVector polygon delimiting the region of 'rasts' within which to compute the models; see ?getRegion for suggestions}

\item{res}{spatial resolution (pixel size), if larger than the original variable raster layers', in which case terra::aggregate() is used}
}
\description{
Get everything (package wrapper)
}
\author{
A. Marcia Barbosa
}
