% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRegion.R
\name{getRegion}
\alias{getRegion}
\title{getRegion}
\usage{
getRegion(occs, type = "mean", prop = 0.5)
}
\arguments{
\item{occs}{species occurrence coordinates (in this order: x, y or LONgitude, LATitude) in geodetic degrees, or an object from which such coordinates can be extracted (e.g. a SpatVector of points)}

\item{type}{character indicating whether to use the "mean" pairwise distance among points (the default), or a fraction of the "width" (minimum diameter) of the points' spatial extent}

\item{prop}{if type="width", proportion of the width to use. Default 0.5}
}
\description{
getRegion
}
