% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modRegion.R
\name{modRegion}
\alias{modRegion}
\title{modRegion}
\usage{
modRegion(occ, type = "mean", prop, map = TRUE)
}
\arguments{
\item{occ}{species occurrence coordinates (in this order: x, y or longitude, latitude), or an object from which coordinates can be extracted (e.g. a SpatVector of points)}

\item{type}{character indicating whether to use the "mean" pairwise distance among points (the default), or a fraction of the "width" (minimum diameter) of the points' spatial extent}

\item{prop}{if type="width", proportion of the width to use, e.g. 0.5}

\item{map}{logical value indicating whether to plot (map) the result}
}
\description{
modRegion
}
