% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varsAvailable.R
\name{varsAvailable}
\alias{varsAvailable}
\title{varsAvailable}
\usage{
varsAvailable(source = NULL)
}
\arguments{
\item{source}{the source database where to check available variables. By default (i.e. if left NULL), all implemented sources are used. Currently "TerraClimate" is the only implemented source.}
}
\value{
character vector
}
\description{
varsAvailable
}
\examples{
varsAvailable()
}
