% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecotrends-package.R
\docType{package}
\name{ecotrends-package}
\alias{ecotrends}
\alias{ecotrends-package}
\title{ecotrends: Temporal Trends in Ecological Niche Models}
\description{
This package computes temporal trends in environmental suitability obtained from ecological niche models, based on a set of species presence point coordinates and predictor variables.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AMBarbosa/ecotrends}
  \item Report bugs at \url{https://github.com/AMBarbosa/ecotrends/issues}
}

}
\author{
\strong{Maintainer}: A. Marcia Barbosa \email{ana.marcia.barbosa@gmail.com}

Authors:
\itemize{
  \item João Alírio
  \item Nuno Garcia
  \item João Campos
  \item Ana Cláudia Teodoro
  \item Lia Barbosa
  \item Isabel Pôças
  \item Salvador Arenas-Castro
  \item Neftalí Sillero
}

}
\keyword{internal}
