% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecotrends-package.R
\docType{package}
\name{ecotrends-package}
\alias{ecotrends}
\alias{ecotrends-package}
\title{ecotrends: Ecological niche model temporal trends}
\description{
Compute temporal ecological niche model trends.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AMBarbosa/ecotrends}
  \item Report bugs at \url{https://github.com/AMBarbosa/ecotrends/issues}
}

}
\author{
\strong{Maintainer}: A. Marcia Barbosa \email{ana.marcia.barbosa@gmail.com} (\href{https://orcid.org/0000-0001-8972-7713}{ORCID})

Authors:
\itemize{
  \item Neftali Sillero
}

}
\keyword{internal}
