% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getImportance.R
\name{getImportance}
\alias{getImportance}
\title{Get variable importance}
\usage{
getImportance(mods, nper = 10, verbosity = 2)
}
\arguments{
\item{mods}{output of \code{\link[=getModels]{getModels()}}.}

\item{nper}{integer value (default 10; increase for more accurate but computationally intensive results) indicating the number of permutations for shuffling each variable.}

\item{verbosity}{integer value indicating the amount of messages to display. The default is 2, for the maximum number of messages available.}
}
\value{
A data frame with the permutation importance of each variable in each model replicate for each year, along with the cross-replicate mean and standard deviation.
}
\description{
This function computes the permutation importance of each variable in each model, by shuffling each variable in turn (a given number of times) and computing the root mean squared difference between the actual model predictions and those obtained with the shuffled variable. Note that "importance" is a vague concept which can be measured in many other different ways.
}
\seealso{
\code{varImportance} in package \pkg{predicts}; \code{bm_VariablesImportance} in package \pkg{biomod2}
}
\author{
A. Marcia Barbosa
}
