% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModels.R
\name{getModels}
\alias{getModels}
\title{Get models}
\usage{
getModels(
  occs,
  rasts,
  region = NULL,
  nbg = 10000,
  nreps = 1,
  collin = TRUE,
  file = NULL,
  verbosity = 2
)
}
\arguments{
\item{occs}{species occurrence coordinates (2 columns in this order: x, y or LONGitude, LATitude) in an object coercible to a data.frame (e.g. a data.frame, matrix, tibble, sf object or SpatVector of points), and in the same coordinate reference system as 'rasts'}

\item{rasts}{(multi-layer) SpatRaster with the variables to use in the models. The layer names should be in the form 'varname_year', e.g. 'tmin_1981', as in the output of \code{\link[=getVariables]{getVariables()}}}

\item{region}{optional SpatExtent or SpatVector polygon delimiting the region of 'rasts' within which to compute the models; see \code{\link[=getRegion]{getRegion()}} for suggestions}

\item{nbg}{integer value indicating the maximum number of background pixels to use in the models. The default is 10,000, or the total number of pixels in the modelling region if that's less.}

\item{nreps}{integer value indicating the number of train-test replicates to compute for evaluating each model. One 1 is implemented currently.}

\item{collin}{logical value indicating whether multicollinearity among the variables should be reduced prior to computing each model. The default is TRUE, in which case the \code{\link[collinear:collinear]{collinear::collinear()}} function is used, with the default values and with the species presences as 'response'.}

\item{file}{optional file name (including path, not including extension) if you want the output list of model objects to be saved on disk. If 'file' already exists in the working directory, models are imported from there.}

\item{verbosity}{integer value indicating the amount of messages to display. The default is 2, for the maximum number of messages available.}
}
\value{
a list of model objects of class \link{maxnet}
}
\description{
This function computes a \code{\link[maxnet:maxnet]{maxnet::maxnet()}} ecological niche model per year, given a set of presence point coordinates and yearly environmental layers.
}
\author{
A. Marcia Barbosa
}
