% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPredictions.R
\name{getPredictions}
\alias{getPredictions}
\title{Get predictions}
\usage{
getPredictions(rasts, mods, file = NULL)
}
\arguments{
\item{rasts}{(multi-layer) SpatRaster containing the variables (with the same names) used in the models}

\item{mods}{list of model objects produced with getModels()}

\item{file}{optional file name (including path, not including extension) if you want the output rasters to be saved on disk}
}
\value{
multi-layer SpatRaster with the predicted values from the variables for each year
}
\description{
Get predictions
}
\author{
A. Marcia Barbosa
}
