% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPredictions.R
\name{getPredictions}
\alias{getPredictions}
\title{Get predictions}
\usage{
getPredictions(
  rasts,
  mods,
  region = NULL,
  type = "cloglog",
  clamp = TRUE,
  file = NULL,
  verbosity = 2
)
}
\arguments{
\item{rasts}{(multi-layer) SpatRaster containing the variables (with the same names) used in the models}

\item{mods}{list of \link{maxnet} model objects, output of \code{\link[=getModels]{getModels()}}}

\item{region}{optional SpatExtent or SpatVector polygon delimiting the region of 'rasts' within which to compute the predictions. The default is NULL, to use the entire extent of 'rasts'. Note that predictions may be unreliable outside the 'region' used for \code{\link[=getModels]{getModels()}}}

\item{type}{character value to pass to \code{\link[=predict]{predict()}} indicating the type of response to compute. Can be "cloglog" (the default), "logistic", "exponential" or "link"}

\item{clamp}{logical value to pass to \code{\link[=predict]{predict()}} indicating whether predictors and features should be restricted to the range seen during model training. Default TRUE}

\item{file}{optional file name (including path, not including extension) if you want the prediction rasters to be saved on disk. If 'file' already exists in the working directory, the rasters are imported from there.}

\item{verbosity}{integer value indicating the amount of messages to display. The default is 2, for the maximum number of messages available.}
}
\value{
multi-layer SpatRaster with the predicted values from the variables for each year
}
\description{
Get predictions
}
\author{
A. Marcia Barbosa
}
