% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRegion.R
\name{getRegion}
\alias{getRegion}
\title{Get region}
\usage{
getRegion(occs, type = "mean", prop = 0.5)
}
\arguments{
\item{occs}{species occurrence coordinates (in this order: x, y or LONgitude, LATitude) in geodetic degrees, or an object from which such coordinates can be extracted (e.g. a SpatVector of points)}

\item{type}{character indicating whether to use the "mean" pairwise distance among points (the default), or a fraction of the "width" (minimum diameter) of the points' spatial extent}

\item{prop}{if type="width", proportion of the width to use. Default 0.5}
}
\description{
This function suggests a region for building an ecological niche model around a given set of species occurrence point coordinates. Mind that this region does not consider survey effort, geographical barriers or other factors that should also be taken into account when delimiting a region for modelling.
}
\seealso{
\code{\link[terra:buffer]{terra::buffer()}}, \code{\link[terra:width]{terra::width()}}
}
\author{
A. Marcia Barbosa
}
