% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTrend.R
\name{getTrend}
\alias{getTrend}
\title{Get ecological trend}
\usage{
getTrend(mods, rasts)
}
\arguments{
\item{mods}{model objects, output of getModels()}

\item{rasts}{output of getVariables(), or another list of variable raster layers, with one multi-layer SpatRaster per year.}
}
\description{
Get ecological trend
}
