% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVariables.R
\name{getVariables}
\alias{getVariables}
\title{Get variables}
\usage{
getVariables(
  source = "TerraClimate",
  vars = varsAvailable(source)$vars,
  years = varsAvailable(source)$years,
  region = c(-180, 180, -90, 90),
  file = NULL
)
}
\arguments{
\item{source}{source to import the raster variables from, e.g. "TerraClimate" (currently the only one implemented)}

\item{vars}{character vector of the names of the variables to be imported. Run varsAvailable() for options. The default is to download all available variables from the specified 'source'. Note that the download can take a long time, especially for many variables, long series of years, and/or large regions.}

\item{years}{year range to get the variables from (e.g. 1979:2013). Note that the download can take a long time for long series of years.}

\item{region}{optional length-four numeric vector (xmin, xmax, ymin, ymax geodetic coordinates in degrees), SpatExtent or SpatVector polygon delimiting the region of the world for which the variables should be downloaded. See ?getRegion for suggestions. The larger the region, the slower the download.}

\item{file}{optional file name (including the path, not including the filename extension) if you want the downloaded rasters to be saved on disk, in which case they are saved as a compressed multi-layer GeoTIFF}
}
\value{
multi-layer SpatRaster
}
\description{
Get variables
}
