% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixelArea.R
\name{pixelArea}
\alias{pixelArea}
\title{Pixel area}
\usage{
pixelArea(r, type = "mean", unit = "m", mask = TRUE, map = TRUE)
}
\arguments{
\item{r}{SpatRaster for which to compute the pixel area}

\item{type}{character value indicating whether the output value should be the "mean" (default) or "centroid" pixel area}

\item{unit}{numeric value indicating the units for the output value: either "m" (default) or "km" squared.}

\item{mask}{logical value (default TRUE) indicating whether to include only the areas of non-NA pixels}

\item{map}{logical value (default TRUE) indicating whether to plot a map}
}
\value{
numeric value
}
\description{
Pixel area
}
